[Keep Moving 29] [14, June, 2018] List the subtasks in devmemo.txt

+++ NgramModel/src/engine/NgramRunEngine.java
@@ -15,6 +15,6 @@ import tokenunit.Tokensequence;
 
 interface NgramRunEngine<K> {
 	public Optional<K> tokenInference(Tokensequence<K> nseq);
-	public float probabilityCalculation(Tokensequence<K> nseq);
+	public double probabilityCalculation(Tokensequence<K> nseq);
 	public void run();
 }
